requirements.txt
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-crispy-forms==2.4
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
mysqlclient==2.2.7
pillow==12.0.0
python-decouple==3.8
python-dotenv==1.1.1
reportlab==4.4.4
requests==2.32.5
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
